This is a re-factoring patch that is a follow-up to the fix for Bug#11873324.
The patch for Bug#11873324 introduced a new member function called
has_guarded_conds() for the JOIN_TAB and the TABLE_REF classes. This
function checks if the JOIN_TAB/TABLE_REF has triggered/guarded
conditions. There exists other places in the optimizer code that do
this check directly. The code for some of these places can be
simplified by using the new member functions. This is implemented in
this patch by replacing this code with calls to has_guarded_conds().

This is a re-factoring patch and should have no changes to existing
behavior.