Fix for bug Bug#12355958 "FAILING ASSERTION: TRX->LOCK.N_ACTIVE_THRS == 1".
The assert that crashes in InnoDB is caused by the ICP code pushing
down an index condition that contains a subquery. When InnoDB
evaluates this index condition the subquery is executed and this results in
a second access to data in InnoDB. InnoDB does not allow a transaction
to have more than one concurrent executing table operation and this 
causes the assert to trigger.

The ICP code should not have pushed down this index condition to InnoDB
since it contains a subquery. The existing ICP code checks for subqueries in
the condition by checking the value of the condition's with_subselect
member. In this case the value for the Item_func condition's with_subselect
is false even though one of the arguments for the Item_func is a subquery.
This situation can occur when we initially create the item tree
for a condition and then later substitutes non-subquery parts of it
with a subquery. In the test case for this bug this occurs when
optimize_conds() replaces a field reference with a subquery due to the
subquery will be const when it is evaluated. This replacement is done
without updating the Item_func node's with_subselect to true.

The fix for this would be easy if it was possible to update the 
with_subselect member locally when doing changes. What makes this
more complex is that this information needs to be propagated 
upwards in the item tree. Since we have no "parent reference" in
the item tree this can not be done locally. In order to be able
to propagate this information upwards in the item tree this patch
extends Item::update_used_tables() to also update
the value of with_subselect when traversing the item tree.