Bug#12608157 - ASSERT IN FIELD_LONG::VAL_INT WHEN USING                MEMORY ENGINE
Item_in_subselect does not have it's own implementation of 
walk(), but rather inherits Item_subselect::walk(). This is a
problem because Item_in_subselect has Item *left_expression 
which Item_subselect does not have.

Due to this, cond->walk() would not walk 
Item_in_subselect::left_expression which resulted in an ASSERT
because the column in left_expression was not marked in the 
read_set. The reason why memory storage engine was required 
to trigger the bug is that with this engine, filesort does 
not use addon_fields. In the case of engines like MyISAM and 
InnoDB, the column not registered in the read set was found 
during preparation of addon fields in register_used_fields().

The fix is to implement Item_in_subselect::walk() which calls
both left_expr->walk() and Item_subselect::walk().