This fix affects pushed queries only.
There was an error in the code setting the T_SCAN_REPEATABLE flag. The
flag was set on the parent of the intended operation rather than on the 
operation itself. This caused scan ranges for the repeatable scan to be 
released too early. This gave wrong results when running a bushy scan where
the scan to be repeated used pruning.