A follow-up for patch that fixed bug #12652385 - "61493: REORDERING COLUMNS TO POSITION FIRST CAN CAUSE DATA TO BE CORRUPTED".
Moved code which is responsible for detecting that ALTER
TABLE can't be carried out using in-place algorithm because
it reorders some columns or replaces some old columns with
a newly added to one place. Introduced a new auxiliary
function for performing this check.

Note that we cannot easily make this check part of
mysql_compare_tables() call as this check is specific to 
normal ALTER TABLE (i.e. assumes that there are old columns 
and new columns). While mysql_compare_tables() is also
used by ALTER TABLE EXCHANGE PARTITION implementation for 
checking that table and partitioned table has identical
structure (i.e. in situation when there are no old and
new columns).