Bug#12616131 - JCL: NULL VS DATE + TWICE AS MANY ROWS RETURNED WHEN JCL>=7
Consider a query of the form

SELECT * FROM t1 LEFT JOIN t2 on t1_col = t2_col

If the BKA UNIQUE algorithm is used on table t2, rows from t1 are read
into the join buffer and each row's "key" (i.e., the columns used in
the join) are then inserted into a hash table. When the buffer is
either full or there are no more rows in t1, join matches are found in
t2 using the key values in the hash table.

The key value that is inserted into the hash table is created by
calling cp_buffer_from_ref(). This effectively formats the key
as a t2_col and copies the necessary info from JOIN_REF.

The problem in this bug was that in the case where t1_col is nullable
but t2_col is not, the information about NULL-values in t1_col was
lost in cp_buffer_from_ref(). Thus, if t1_col had a value NULL, key
would not be NULL but 0 (not necessarily reliably). The result was
that t1 rows with a t1_col value of NULL would be joined with rows in
t2 with t2_col=0.

The fix is to not add NULL-valued keys into the hash table if the join
condition is '='. This also saves work since these keys cannot result
in join matches anyway.

Note 1: The information about NULL values in t1_col is not lost as
long as t2_col is also nullable.

Note 2: If the join condition is <=>, the join correctness is handled
by a triggered condition added in pushdown_on_conditions()