Consider a query containing a predicate of the form
  left_expr IN/SOME/ANY/ALL (subquery)

If any part of left_expr is UNKNOWN, then this predicate shall
evaluate to: 

  - UNKNOWN if the subquery returns rows after disabling 
    subquery conditions relying on the NULL outer values
  - false otherwise

For normal subqueries, this is handled by disabling any subquery
predicates relying on UNKNOWN values in left_expr and then
evaluate the subquery. If the subquery returns at least one row
the predicate evaluates to UNKNOWN.

However, when BKA is used to resolve the subquery, the values in
left_expr are used to lookup rows that are join matches. BKA
cannot disable parts of the key when doing this lookup and can
therefore not be used in such cases.

The problem was that BKA was used in cases where left_expr might
have UNKNOWN values. The fix is to detect and disable BKA in such
cases.