Build failed after upgrade of GCC to 4.6.1
After upgrading to gcc 4.6.1, the build failed with:
 error: cast to pointer from integer of different size 
       [-Werror=int-to-pointer-cast]

The change log for gcc (http://gcc.gnu.org/gcc-4.6/changes.html) 
says:
"GCC now warns by default when casting integers to larger pointer
types. These warnings can be disabled with the option 
-Wno-int-to-pointer-cast, which is now also available in C++"

The line has been removed as it is a remnant from old code, and
the result of the assignment is no longer used.