Bug#11938817 ALTER BEHAVIOR DIFFERENT THEN DOCUMENTED
The problem was that "ALTER TABLE t1" did nothing while 
"ALTER TABLE t1 ENGINE = x" rebuilt the table (given that
t1 was using engine x to start with). Even if both statements
logically do not change to the table.

After discussing the issue with support, it was decided that we
cannot easily change the behavior of "ALTER TABLE t1 ENGINE = x"
as users and external tools depend on the current behavior.

This patch adds a code comment explaining current behavior,
as well as a regression test documenting the behavior.