WL#5535 - Separate tablespace(s) for the InnoDB undo log
Allow users to move the additional UNDO logs to separate tablespaces. For this
we introduce 3 new parameters and rename an existing parameter.
 
    1. innodb_undo_directory (string)
       Path to rollback tablespace directory, default is ".". Value can be
       any file path. If it is not absolue then create the rollback tablespaces
       relative to the server CWD.
 
    2. innodb_undo_tablespaces (ulint)
       We allow 1-1 mapping and 1-N mapping between rollback segments and
       tablespaces. ie. we can create multiple segments in one or more
       tablespaces.
 
    3. innodb_undo_logs (ulint)
       Default is TRX_SYS_NRSEGS, if value is greater than current rollback
       segments then create the additional rollback log segments. If
       innodb_undo_tablespaces > 0 then they will be created outside ibdata1
       otherwise in ibdata1.


Rename innodb_rollback_segments to innodb_undo_segments so that names are
consistent.

Refuse to startup if innodb_undo_tablespaces doesn't match what InnoDB can
access.

Change the os0file.c create mode parameter from #defines to an enum.

Reduce the scope of srv_file_per_table to table creation code in dict0crea.c.
Rather than access the (global) variable in os0file.c we note the setting and
pass the OS_FILE_ON_ERROR_NO_EXIT flag down the call stack.

Users cannot add new UNDO tablespaces after database creation. Creating the
UNDO tablespaces is a one time affair. The segments however can be added
to existing UNDO tablespaces or to the system tablespace if no UNDO tablespaces
exist.

NOTE: Databases created with the above options enabled cannot be opened by
previous versions of InnoDB. The file format is not an issue, the problem is
that there are several assumptions in the code that will not hold. To use
this feature the config vaiable innodb_undo_tablespaces must be set to a
value > 0 when a new database is created. Users cannot add new UNDO
tablespaces after database creation. Creating the UNDO tablespaces is a
one time affair. The segments however can be added to existing UNDO
tablespaces or to the system tablespace if no UNDO tablespaces exist.
    
HotBackup ramifications: The new variables need to be defined and set
correctly in HB code.
    
By default this feature is disabled.

rb://561 Approved by Marko