Bug#11890269 - PERFSCHEMA.TABLE_AGGREGATE_THREAD_2U_3T FAILS OCCASIONALLY ON PB2
This is a test script bug only, the server code is not affected.

Before this fix, tests for table io and table lock aggregation would fail
sporadically.

Failures occurred in the output of this prepared select:

prepare dump_thread from
  "select event_name, count_star
  from performance_schema.events_waits_summary_by_thread_by_event_name
  where event_name in
  (\'wait/io/table/sql/handler\',
   \'wait/lock/table/sql/handler\')
  and thread_id in
  (select thread_id from performance_schema.threads
   where processlist_user in (\'user1\', \'user2\', \'user3\', \'user4\'))
  order by thread_id, event_name;";

This statement performs a sub query and a file sort on performance schema
tables. Because the content of the performance schema tables is volatile,
this kind of statement is not reliable.

Also, the test output was difficult to interpret, since which thread_id /
user name corresponds to which statistics was not printed.

This fix replaced the unreliable "dump_thread" prepared statement by a stored procedure,
which performs selects without sub queries, and prints more detailed output.

All the tests result files have been adjusted accordingly.