Fix for Bug#12601961 SEGFAULT IN HANDLER::COMPARE_KEY2.
The crash in InnoDB's ICP implementation occured due to the
handler's end_range member variable having an invalid value.
As part of the ICP evaluation in InnoDB the code will check
that the record is within the end_range of the query. In the
case where this leads to a crash the end_range was not 
initialized and the code could then reference a non-valid
memory area leading to the segmentation fault (it could probably
also lead to a wrong result if the "random end_range" decided
that the actual record was outside the valid record range).

The case where ICP can be used without having a proper
end_range defined is for "const" queries that will at most
access one record. In the case where this occurs the
table is read from join_read_const() (in sql_select.cc) which
retrieves data using handler::ha_index_read_idx_map(). 
Since only max one record will be read, ha_index_read_idx_map()
does not define an end_range for the handler. In this case
end_range might be undefined when a pushed index condition
is evaluated. 

The fix for this problem is to intialized end_range in the
handler's constructor. The patch also adds an assert to
ha_index_read_idx_map() to detect if it ever is called
with an end_range value different from NULL.

Without this fix the test case will either crash or pass
depending on the value the unitialized end_range get. Running
with valgrind it should produce a test failure.