Refactoring as part of preparing SPJ results to be double (multi...?) buffered.
'class SharedFragStack' which was intended as a container for fragments
which has received a complete batch of rows is removed. 
It is replaced with a set of (mutex protected) methods in 
'class NdbRootFragment' and OrderedFragSet::prepareMoreResults()

The rational for this refactoring is that the SharedFragStack container
wasn't a particular good fit to represent the state when *multiple*
batches of (double buffered) result set may be available.