Bug #11766001: 59026: ALLOW MULTIPLE --PLUGIN-LOAD OPTIONS
Implemented support for a new command line option :
--plugin-load-add=<comma-separated-name-equals-value-list>
This option takes the same type of arguments that --plugin-load does
and complements --plugin-load (that continues to operate as before) by
appending its argument to the list specified by --plugin-load.
So --plugin-load can be considered a composite option consisting of 
resetting the plugin load list and then calling --plugin-load-add to process
the argument.
Note that the order in which you specify --plugin-load and --plugin-load-add 
is important : "--plugin-load=x --plugin-load-add=y" will be equivalent to
"--plugin-load=x,y" whereas "--plugin-load-add=y --plugin-load=x" will be 
equivalent to "plugin-load=x".

Incompatible change : the --help --verbose command will no longer print the 
--plugin-load variable's values (as it doesn't have one). Otherwise both --plugin-load 
and --plugin-load-add are mentioned in it.
Added few test cases to test different combinations.
Added an extension to mysql-test-run.pl to add variables for --plugin-load-add similarly
to what it does for --plugin-load.
Moved the includes on top of the files.