WL#5986: Stored programs: make SQL-handlers closer to the standard).
The patch changes the way how SQL handlers are found and activated.

We used to search handlers recursively (from the current scope up to the top
one). That was needed because a search for handlers was done right when an SQL
condition occurs. After patch for Bug 23032 the search for SQL handlers happens
after SQL statement completes.

That means, there is no need to do a recursive search now:
  - if there are warnings, they are handled by SQL handlers within
    the current scope only;
  - if there is an error, it is propagated to the outer scope.