Regenerated test result fro ndb_statistics1 with notes.
Three query plans changed but is equally good as the previous.
A test related to bug #59519 seems broken even before the
previus commit. The query plan should have >1 row in last join.

In the same query the index for (I,J) should be choosen but
in this case the resulting index (J,K) is equally good.