BUG#11878394 - MYSQLD_SAFE TEST FOR ALREADY RUNNING PROCESS                FAILS, CAUSING CLOBBERED SOCKET
A check for running mysqld instances was failing in mysqld_safe
because of an incorrect shell command generated by cmake. The
problem is that cmake retains the '\' of escaped double-quote,
so the generated mysqld_safe script's command contained \" in
the shell command, and hence the failure. 

Fixed the command in scripts\CMakeLists.txt.