BUG#11766826: TEST RPL.RPL_SLAVE_LOAD_REMOVE_TMPFILE AGAIN FAILS WITH --MEM
The test case rpl_slave_load_remove_tmpfile checks if the 
slave fails gracefully when the temporary file used for
processing LOAD DATA is removed when it is about to be used.

Now... Fix for BUG 11765751 changed the server behavior due to
security reasons. The logic that must be enforced is:
1. normalize the string 2. check against pattern 3. check
file. However, in the case of broken symlinks the server/slave
will not be able to normalize the path, thus will end up
mathching normalized against non-normalized paths. Eventually, it
will end up with an error stating that LOAD FILE operation is not
allowed because it does not match the slave-load-tmpdir pattern.

We fix the test case by adding the error code 1290 to the list of
expected errors.