Fix for Bug#11766559 "59696: Optimizer fails to move WHERE condition on JOIN column when joining with a view".
When optimizing a JOIN of two tables the optimizer tries to move WHERE
conditions on the join column(s) to the first table in the JOIN in
order to reduce the number of records it has to read. If the second
operand of the JOIN is a view that is identical to the underlying
table the optimizer should be able to perform the same optimization.
In this case it fails to move the WHERE condition on the join column
to the first table.

The cause for this is that when the second operand is a view then
the conditions contain Item_ref objects that reference the real item
field objects. The existing code fails to evaluate these since it
only evaluates the Item_ref object and fails to look at the real
objects. As a result it fails to detect that the equals predicate on
the second join operand can be substituted by a corresponding equals
predicate on the first operand.

There are two places where we fail to handle Item_ref correctly and that
causes the optimizer to fail to do this optimization:

1. In optimize_cond() we call build_equal_items() which again calls
build_equal_items_for_cond() where we in Item_func::compile() fail to
evaluate the Item::equal_fields_propagator() for the condition on the
view operand. As a result the field used in the condition for the view
is not considered as an equal field to the join column of the first
table when this later is evaluated in make_join_statistics().

2. Later in JOIN::optimize() we call substitute_for_best_equal_field()
which uses Item_func::transform() to select the best condition of the
tables. Item_func::transform() fails to evaluate the different
conditions that are equal due to these being stored in a Item_ref
object.

The fix for these problems is to implement compile() and transform()
functions for the Item_ref class that compile/transform both the
Item_ref object itself and the item object referenced by the 'ref'
pointer. So when Item_func::compile() and Item_func::transform() 
evaluate Item_ref object this fix ensures that both the Item_ref 
object and the object it references are evaluated.