Fix for bug #11766767 - "59957: VIEW USING MERGE PERMISSIONS IN MULTI-TABLE UPDATE".
Multi-update required UPDATE privilege on a view which was
only read by it if this view was processed using MERGE 
algorithm. At the same time in this situation SELECT privilege 
was enough for a base table or a view which was processed using 
TEMPTABLE algorithm. 

The problem stemmed from the fact that code in 
mysql_multi_update_prepare() used TABLE_LIST::updating attribute
in order to determine which privilege UPDATE or SELECT require
on particular table in the main select of multi-update. Since
this attribute was not correctly set in some cases (including
the above case with a mergeable view) this privilege check was
not working correctly. There were also other problems with
the approach used (see comment for sql_update.cc for details).

This patch solves the problem by re-implementing this privilege
check in such way that it no longer relies on TABLE_LIST::updating
attribute. Instead it directly uses bitmap of updated tables to 
determine which of tables from the multi-update's main select were 
updated.

Issues with replication filtering and handling of --read-only
option which are caused by incorrectly set TABLE_LIST::updating
member were reported as bugs 12979999 and 12980522.
They are to be investigated and fixed separately.