Bug#12859831 IGNORED ERRORS ARE CONSIDERED DELETE OPERETIONS CRASHING SERVER WITH AN ASSERT
Recent code changes as part of read-before-write removal assert 
that the number of operations that fail due to non-existing keys should
be <= the number of defined operations *of a single type*.  This works
well for the normal SQL client handler call patterns, but the Slave SQL
thread defines operations of multiple types in batches together, which can
cause these assertions to fail.

No testcase yet.

Not applied to 6.3.