WL#5986: Stored programs: make SQL-handlers closer to the standard).
This patch changes how conditions are removed from the diagnostics
area when handlers are activated.

Before, the handled condition was removed when the handler was
activated. Now, any active conditions are marked when the
handler is activated and then later removed when the handler
exits. This allows diagnostics statements (e.g. SHOW WARNINGS)
to inspect these conditions inside the handler. (Normal diagnostic
area rules still apply, so the diagnostic statement will have
to be the first statement in the handler.)

Note that this does not apply to RESIGNAL with a condition value.
In this case, the handled conditions will be kept. (This matches
existing behavior.)