Fix for Bug#12976163 "CRASH IN INDEX CONDITION PUSHDOWN CODE AGAINST A MYISAM TABLE"
The crash in ha_myisam::idx_cond_push() occured due to the server was
trying to push an index condition down to MyISAM on a temporary table.
This temporary table was going to store data for a derived table. 
The reason for the crash was that the handler object was not completely
initialized due to the underlying temporary table not being created yet.
This situation was introduced by WL#5274 where we postpone creation of
temporary tables until execution.

The crash occured when ha_myisam::idx_cond_push() tried to access
handler->table->key_info when was handler->table was NULL. The fix is to 
use handler->table_share instead of handler->table.