BUG#12838171:  51VS56: TRANSFORMED IN()+SUBQ QUERY PRODUCES                EMPTY RESULT ON 5.6, 1 ROW ON 5.1
Consider a query of the form:
      
SELECT ... 
WHERE outer_value_list IN (SELECT inner_value_list ... 
                           WHERE colX IS NULL)
      
If outer_value_list contains a NULL value, the IN predicate 
shall evaluate to FALSE or UNKNOWN (depending on whether or 
not the inner query returns rows). In any case the predicate 
is not satisfied.

If IN->EXISTS transformation is chosen and there is an index 
that covers both inner_value_list and colX, the inner query 
is resolved using subquery_uniquesubquery_engine. This engine 
had a shortcut when copying the key part values used for 
index lookup: if one of the key parts had a NULL value and the
IN subquery was "top level", it would return "predicate not 
satisfied".

This shortcut would have been OK if the NULL value came from 
the outer_value_list, but in this bug colX has the NULL 
value. Thus, the predicate should return true whereas it 
returned false due to the shortcut.

subselect_[unique|index]subquery_engine::exec() share the code
that has the mentioned shortcut 
(subselect_uniquesubquery_engine::copy_ref_key()). Both
methods are called only from Item_in_optimizer::val_int() 
which has the same shortcut. This means that it will never be 
the case that an outer value is NULL for a top level item in 
copy_ref_key(). The fix is therefore to remove the shortcut.