WL#5986: Stored programs: Make condition handlers closer to the standard.
Fix assertion failure if a variable declaration raises a warning
and a warning handler is declared. The assertion was triggered
because the parser thought the handler should be activated, yet
at runtime the handler declaration had not been processed yet.

This patch fixes the assertion by not activating the handler in
such cases (this matches the current 5.5/5.6 behavior).

Test case added to sp-error.test.