Bug#13068506: QUERY WITH GROUP BY ON NON-AGGR COLUMN RETURNS               WRONG RESULT
Consider the query:

SELECT (
  SELECT MAX( t2.i2 )
  FROM t3 RIGHT JOIN t2 ON ( t2.i3 = 2 )
  WHERE t2.i3 <> t1.i1
) AS field1
FROM t1
GROUP BY field1;

In simplify_joins(), the optimizer checks if the outer join can
be transformed to an inner join, and if the ON conditions 
can be transformed to WHERE conditions. This can be done if 
there are conditions that reject null-extended rows (extensions
to t3 in the query above)

To check that the inner tables does have a join match for all
qualifying outer rows, not_null_tables() is called on the 
conditions and if all inner tables will produce rows the join
is transformed (see simplify_joins()):

if (conds)
  not_null_tables= conds->not_null_tables();
...
if (!table->outer_join || (used_tables & not_null_tables))
{
  <transformation happens here>
}

In the query above, t3 has table->map=1 and t2 has 
table->map=2. Thus, when processing t2 of the subselect,
used_tables=1. The WHERE condition is not null-rejecting for
table t3, so not_null_tables shold not have bit 0 set. But 
it has.

Item_outer_ref inherits not_null_tables() from Item_ref. 
Table t1 has table->map=1 in the outer select, so when 
not_null_tables() is called on the t1.i1 part of the <> 
condition, the map of t1 is returned. The bug is that
Item_outer_ref::not_null_tables() returns the map of the outer
reference in the first place. It should return 0.