Resurected code which implemented the new MRR interface from branch mysql-6.0-telco-7.0.
Several fixes was required in order to make that code work as a looong time
has passed since it was originally developed:

 - Addapted fix for Bug#57481:'multi range read' may fail to close completed NdbScanOperations'
 - Start transactions on demand when MRR operations are created.
 - Corrected bug in calculating size of MRR buffer to allocate.
 - Addapted code to correctly count pruned/sorted & plain scans in MRR
 - Addapted code to handle user specified partition.

SPJ has not yet been integrated into the new MRR code - instead we will (temporary)
use the default MRR impl when a join is pushed.

Old MRR code is still present as a reference. (Plan to remove that in a later push)