Fix Bug#12661768 UPDATE IGNORE CRASHES SERVER IF TABLE IS INNODB AND IT IS PARENT FOR OTHER ONE
When "UPDATE t1 SET c = 1" is prevented by a duplicate entry in t2 (where
there is a FK on t1 and t2 and ON UPDATE CASCADE), then InnoDB returns the
number of the offending index from t2, but the MySQL part of the code
assumes this is the Nth index from t1. If t1 contains less indexes, then a
crash may occur, or if the returned number is less than the number of the
indexes - then a bogus data is printed.

Since MySQL is not aware of the second table, the option here is to return
its name and the index name from InnoDB to MySQL instead of just a bare
number (designating the number of the index).

Approved by:	Jimmy, Marko, Jon Olav Hauglid (rb://768)