Bug#13251277 CHECK_GRANT SHOULDN'T LOCK/UNLOCK LOCK_GRANT              WITHOUT NEEDING TO
The problem was that check_table_access() was called for each
stored routine instruction even if the instruction did not use
any tables. This is a performance problem since
check_table_access() causes the LOCK_grant rw lock to be taken
(by calling check_grant()). For example, it caused up to 20 %
slowdown in a simple stored routine calculating Fibonacci's
numbers.

This patch fixes the problem by changing the code so that
check_table_access() is only called for stored routine
instructions that actually use tables.

The patch doesn't contain a test case since it is hard to test
for performance improvements in our test framework.