Bug#12997905: VALGRIND: SYSCALL PARAM PWRITE64(BUF)               POINTS TO UNINITIALISED BYTE(S)
When a record is inserted into a MyISAM temporary table using 
static write, the entire content of the record[0] buffer is 
written to the table which resides on disk. If one or more of
the columns is a varchar, the trailing space may be 
uninitialized and even contain secrets. Valgrind barked because
of this.

The fix is to reset the record buffer for MyISAM temporary 
tables in addition to restricting the copying of fields from
one table to another to the bytes actually used by the varchar.

Bonus: Restricting Copy_field to copy only used varchar bytes 
instead of the full column length also has a performance 
benefit.