Fix for better load balancing SPJ operations, and reduce exec. latency :
Add a deterministic skew in the fragmentlist such that it is rotated 
one place for each SPJ request. This has the effect that a non parallel
scan is not started on the same subset of fragments by all
SPJ executors. Which will result in a better loadbalancing and
reduced latency.