BUG#13247871 - ER_EVENT_CANNOT_ALTER_IN_THE_PAST HAS WRONG MESSAGE
BACKGROUND:
ALTER EVENT, produces the *error* ER_EVENT_CANNOT_ALTER_IN_THE_PAST, 
but the message is wrong. The message string is the exact 
same as for the CREATE EVENT warning: "Event execution
time is in the past and ON COMPLETION NOT PRESERVE is set.
The event was dropped immediately after creation."

FIX:
The error message for ER_EVENT_CANNOT_ALTER_IN_THE_PAST has
been updated and says like this 
"Event execution time is in the past and ON COMPLETION NOT PRESERVE 
is set.The event was not changed. Specify a time in the future."