WL#4143
Fixing a -Werror on behalf of Alfranio (slave_was_killed= 0).

In addition, deployed an error message when get_password
fails to retrieve a password when attempting to connect to the
master and made the function return immediately. Before, there
could be the case that the slave failed to retrieve the password
and later assumed that the connection was established, when 
indeed it was not.