Bug #13357607 INNODB: WRONG ERROR ON STARTUP AFTER CREATING A COMPRESSED TABLE
The patch for WL#5873 - Table Flags Refactor contained a bug in
dict_sys_tables_type_validate() which validates the SYS_TABLES.TYPE field
for each table.  It reported a valid zip_ssize as invalid.  This caused
compressed tables to not be opened on startup.  Surprisingly, there is no MTR
test that caught this. So this patch adds a simple innodb-restart.test and
fixes the bad condition.

Approved by Jimmy Yang in http://bur03.no.oracle.com/rb/r/805/