BUG#11761014 - 53466: END_CONNECTION() SHOULD REFERENCE GLOBAL_SYSTEM_VARIABLES.LOG_WARNINGS
BACKGROUND:
end_connection prints out messages like this, depending on *session*
log_warnings value:
Aborted connection 2 to db: 'unconnected' user: 'root' 
host: '' (Got timeout reading communication packets).
This is wrong because it allows a minimally privileged user
to write copious amounts of errors to the .err file, unless 
he is restricted by the number of MAX_CONNECTIONS_PER_HOUR 
in his user account.

FIX:
"Aborted connection..." messages are now printed based on 
the *global* log_warnings value. Since this was the only real
remaining use of the session value of log_warnings, the 
log_warnings system variable has now been made to a global 
only variable. This is a change of behavior and must be 
reflected in the documentation.