ndb - new template for intrusive hash table, intended to the one
There was several variants of DLHashTable such as DLHashTable2, KeyTable,
KeyTable2 and KeyTable2c.  In time this new template should replace them
all.

The objective to introduce it now is to simplify the use of two hash-tables
on the same object, now needed for DictObject.

The M in DLMHashTable can mean Meta or Methods, since the template takes
a Meta/Methods-class as parameter to tell the hash table how to get
next/prev link, calculate hash-value and determine equalness of objects.

The default Meta/method-class assumes the same interface that is already
used in classes using DLHashTable.

DLMHashTable logic are copied from DLHashTable.  Direct references like
obj->nextHash are replaced with M::nextHash(i*obj) and the like for 
prevHash, hashValue() and equal().