Bug#12876523: VALGRIND FAILURE IN COMPRESSION TESTS
The valgrind issue happens in the zlib package,
and according to the zlib FAQ, it is intentional for
performance reasons (http://www.zlib.net/zlib_faq.html#faq36).

The current valgrind suppression file contains several call
stacks for valgrind issues with deflate().  However, all the
existing stacks differ from the one reported here at the
4th level.

I assume the failing stack trace must be due to inlining
in optimized builds, but I have not been able to reproduce
it on my available platforms.  However, I see no reason that
the suppression rules should distinguish on which MySQL
function call deflate() when there is a known problem with
how deflate() works.  By limiting the stack in the suppression
file to 3 levels, we can also reduce the number of suppression
rules for deflate().